feat: implement modal for fetched post

+++ src/hooks/useLocalStorage.ts
@@ -1,19 +1,21 @@
 import { useState } from 'react';
-import { deleteFromLS, getFromLS, setToLS } from 'utils/index';
+import { deleteFromLS, getFromLS, setToLS } from 'utils/helpers';
 
 function useLocalStorage(key: string) {
   const [storedValuesArray, setStoredValue] = useState(getFromLS(key) || []);
 
-  const updateStoredValuesArray = (text = '', isDelete = false) => {
-    if (!isDelete) {
-      const arrayValue = getFromLS(key);
+  const updateStoredValuesArray = (text: string, isDelete = false) => {
+    if (text !== '') {
+      if (!isDelete && text !== '') {
+        const arrayValue = getFromLS(key);
 
-      if (arrayValue.indexOf(text) === -1) {
-        arrayValue.push(text);
-        setStoredValue(setToLS(key, arrayValue));
+        if (arrayValue.indexOf(text.toLowerCase()) === -1) {
+          arrayValue.push(text);
+          setStoredValue(setToLS(key, arrayValue));
+        }
+      } else {
+        setStoredValue(deleteFromLS(key, text));
       }
-    } else {
-      setStoredValue(deleteFromLS(key, text));
     }
   };
